requirements.txt
flask 
flask_sqlalchemy 
flask-bcrypt 
flask-jwt-extended
datetime
logging
sqlalchemy.exc
functools